requirements.txt
PyQt5>=5.15.0
Pillow>=9.0.0
numpy>=1.21.0
opencv-python>=4.5.0
scikit-image>=0.18.0
psutil>=5.8.0  # For memory management